:recycle: refactory: jogando todas as chamadas de função para a pasta hooks
